training/requirements.txt
# Install PyTorch BEFORE this file — pick the build for your hardware:
#
#   CPU only:
#     pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#
#   NVIDIA GPU (CUDA 12.1):
#     pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
#   NVIDIA GPU (CUDA 11.8):
#     pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
#   AMD GPU — ROCm 6.0 (Linux only):
#     pip install torch torchvision --index-url https://download.pytorch.org/whl/rocm6.0
#
# Python 3.10–3.13 required (3.14+ is incompatible with PyTorch).

numpy>=1.26
opencv-python>=4.9
pyyaml>=6.0
tqdm>=4.66
segmentation-models-pytorch>=0.3.3
timm>=0.9.0
